divide_abstract_queue
Move `AbstractPacketsQueue` to a separate file.
Optimize queue's API.
